# Reduce overhead of transform processing when no transforms are present in a drawable