As noticed by @smoogipoo during multiplayer spectator profiling, `updateTransforms` can become quite expensive as it is called millions of times. Even in cases where there are actually not transforms on drawables.

Before:

|                          Method |     Mean |   Error |  StdDev | Allocated |
|-------------------------------- |---------:|--------:|--------:|----------:|
| UpdateTransformsWithNonePresent | 101.2 us | 0.31 us | 0.25 us |         - |
| UpdateTransformsWithManyPresent | 383.3 us | 2.21 us | 2.07 us |       1 B |

After:

|                          Method |       Mean |     Error |    StdDev | Allocated |
|-------------------------------- |-----------:|----------:|----------:|----------:|
| UpdateTransformsWithNonePresent |   5.091 us | 0.0303 us | 0.0269 us |         - |
| UpdateTransformsWithManyPresent | 353.660 us | 6.8759 us | 9.6390 us |       1 B |

Testing on 30s of the same beatmap:

![WindowsTerminal_2022-07-01_03-39-08](https://user-images.githubusercontent.com/191335/176754217-e6a712a8-db77-40ea-b178-a875ac774842.png)

Note that this will have a larger effect on multiplayer spectator than what is seen in these profiling results.

There's a few other similar cases where I think we need to similarly shortcut to avoid every-frame overheads (`CheckChildLife` is another seemingly expensive / often visited method). Here's a list of function call counts during autoplay (take the timings with a grain of salt as this was using tracing collection):

![unknown](https://user-images.githubusercontent.com/191335/176754800-9398c1ab-728b-495c-a668-fbf70606deca.png)